refactoring rel deck cards

+++ src/main/java/com/naicson/yugioh/service/interfaces/CardDetailService.java
@@ -47,7 +47,7 @@ public interface CardDetailService {
 
 	Page<Card> searchCardDetailed(List<SearchCriteria> criterias, String join, Pageable pageable);
 
-	List<RelDeckCards> findAllRelDeckCardsByCardNumber(Long cardNumber);
+	List<RelDeckCards> findAllRelDeckCardsByCardNumber(Integer cardNumber);
 	
 	List<Long> findCardsNotRegistered(List<Long> cardsNumber);
